Feat/custom events (#31)
* feat: implments date and time scheduled function

* feat: adds custom event registry

* feat: adds custom event class

* feat: implements custom event registry into khome core

* feat: adds onCustomEvent function to register callback for event

* feat: emits custom event with event data

* renames dto to prevent name clashes with custom events

* feat: adds custom event subscription

* feat: converts functional service calls into classes

* fix: removes no workin event emitter

* refactor: makes brightnessPct type double instead int

* refactor: makes service data builder function nullable

* feat: allows calling onCustomEvent with custom event type instead of a name as string

+++ src/main/kotlin/khome/calling/Light.kt
@@ -30,7 +30,7 @@ data class LightData(
     var kelvin: Int? = null,
     var whiteValue: Int? = null,
     var brightness: Int? = null,
-    var brightnessPct: Int? = null,
+    var brightnessPct: Double? = null,
     var profile: String? = null,
     var flash: String? = null,
     var effect: String? = null
